+++ Automation/src/test/java/cucumberOptions/TestRunner.java
@@ -7,9 +7,15 @@ import cucumber.api.junit.Cucumber;
 
 @RunWith(Cucumber.class)
 @CucumberOptions(
-		features = "src/test/java/features"
-		,glue={"stepDefinition"}
+		features = "src/test/java/features/Login.feature"
+		,glue={"com.pb.stepDefination.stepdefinations"}
 		)
+
+//@RunWith(Cucumber.class)@CucumberOptions(plugin = {
+//	    "pretty", "json:target/Open-Google-WebSite.json"},
+//	   features = {"src/test/java/features/Login.feature"},
+//	   glue = {"com.java.StepDefination"})
+
 public class TestRunner {
 
 }
